More Fixes, close to v0.4.5
Lots of refactoring, and finally took care of that pesky rendering issue of #29. The map is still locked at being 3200x1800 right now, but it's dynamic so I can modify it at any time easier. Server/Client loops run faster now too, which is nice!

Every class is in it's own file now, so finally I dont have to scroll through all that code to find the little bit that I need (should done this sooner!). Also added a SPECIAL_THANKS.md file which lists who assisted me and how. Next up: Fix my selection code!
