+++ SPECIAL_THANKS.md
@@ -1,0 +1,62 @@
+# SPECIAL THANKS:
+Sorted in no particular order other than type of assistance.
+
+## Programming Assistance:
+### Seth Berrier:
+**GitHub:** [https://github.com/olliebrown](https://github.com/olliebrown)
+**Assisted with:**
+- Setting up the project
+- Debugging various issues
+- Motivation
+- Concept for GameServer.updateGrid()
+
+### RainingChain:
+**Website:** [https://rainingchain.com/](https://rainingchain.com/)
+**YouTube Tutorial:** [Making Multiplayer HTML5 Game](https://www.youtube.com/watch?v=PfSwUOBL1YQ)
+**Assisted with:**
+- Original idea to make an .io game
+- Base code/structure before swap to ES6
+- Motivation
+
+### Ruben Torres:
+**GitHub:** [https://github.com/rubentd](https://github.com/rubentd)
+**Tanks Repo:** [https://github.com/rubentd/tanks](https://github.com/rubentd/tanks)
+**Assisted with:**
+- Idea to use classes, which led me to switch to ES6
+
+### Gustavo Carvalho:
+**Stack Overflow Answer:** [Here](http://stackoverflow.com/questions/16919601/html5-canvas-camera-viewport-how-to-actally-do-it)
+**Assisted with:**
+- Created the idea and original code for the camera viewport system
+
+### Gregory M:
+**Stack Overflow Answer:** [Here](http://stackoverflow.com/questions/12885001/how-to-get-a-sprite-to-rotate-to-face-the-mouse-position-using-html5-canvas)
+**Assisted with:**
+- Getting gun to rotate correctly towards the mouse
+
+### Cole Love:
+**Assisted with:**
+- Getting gun to rotate correctly towards the mouse
+
+## Art Assistance:
+### Clinton Bone:
+**Website:** [http://clintbone.com/](http://clintbone.com/)
+**Assisted with:**
+- Made grid background
+- Made all other currently unused art assets
+- Testing
+
+## Testing Assistance:
+### Marko Vehmas:
+- **GitHub:** [https://github.com/Dorover](https://github.com/Dorover)
+
+### Reed Kendall:
+- **GitHub:** [https://github.com/ReedKendall](https://github.com/ReedKendall)
+
+### Valmay Green:
+- **Website:** [https://valmaygreen.wixsite.com/valmaygreenportfolio](https://valmaygreen.wixsite.com/valmaygreenportfolio)
+- **Twitter:** [https://twitter.com/ValmayGreen](https://twitter.com/ValmayGreen)
+
+### Tony Dols
+
+### Alex Meyer
