+++ client/views/index.ejs
@@ -10,6 +10,7 @@
        <span class="input-group-btn">
          <input id="play" class="btn btn-primary" type="submit" value="Join the battle">
        </span>
+       <p id="errorMessage" style="display:none;"><br>Error: Name too long, limit it to 9 characters.</p>
      </div>
   </div>
